Create NativeOptimizations plugin
