+++ Apps/Playground/CMakeLists.txt
@@ -114,6 +114,7 @@ target_link_to_dependencies(Playground
     PRIVATE NativeCapture
     PRIVATE ChromeDevTools
     PRIVATE NativeEngine
+    PRIVATE NativeOptimizations
     PRIVATE Console
     PRIVATE Window
     PRIVATE ScriptLoader
