+++ Plugins/NativeOptimizations/Include/Babylon/Plugins/NativeOptimizations.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <napi/env.h>
+
+namespace Babylon::Plugins::NativeOptimizations
+{
+    void Initialize(Napi::Env env);
+}
